chore: Best & New Curation Page를 curation 폴더로 위치
